+++ lib/base.js
@@ -53,26 +53,6 @@ class Base {
   }
 
   /**
-   * Get next page from a previous call.
-   * @private
-   * @param {Object} pageInfo Pages information from a previous call.
-   * @returns {Promise} Returns success if intercom is disabled, otherwise
-   *   returns results of call to intercom client.
-   */
-  nextPage (pageInfo) {
-    return Promise.resolve()
-      .then(() => {
-        if (!Util.canUseIntercom()) { return }
-
-        if (!this.client) {
-          throw new Error(`Base._wrap: invalid ${this.name} client`)
-        }
-
-        return this.baseClient.nextPage(pageInfo).get('body')
-      })
-  }
-
-  /**
    * List Objects. This is pretty universal.
    * @return {Promise} Resolves with the body of the list call for the type.
    */
