+++ lib/orion.js
@@ -2,6 +2,7 @@
 
 const Companies = require('./companies')
 const Intercom = require('intercom-client')
+const Promise = require('bluebird')
 const Tags = require('./tags')
 const Users = require('./users')
 const Util = require('./util')
@@ -36,6 +37,26 @@ class Orion {
       const ClassName = classes[name]
       this[name] = new ClassName(this.intercomClient, name)
     })
+  }
+
+  /**
+   * To maintain compatibility with Intercom's client, this nextPage function
+   * can be used for pagination on any type.
+   * @return {Object} pageInfo Pagination information from a previous request.
+   * @returns {Promise} Returns success if intercom is disabled, otherwise
+   *   returns results of call to intercom client.
+   */
+  nextPage (pageInfo) {
+    return Promise.resolve()
+      .then(() => {
+        if (!Util.canUseIntercom()) { return }
+
+        if (!this.intercomClient) {
+          throw new Error('Orion.nextPage: invalid intercom client')
+        }
+
+        return this.intercomClient.nextPage(pageInfo).get('body')
+      })
   }
 }
