+++ test/orion.js
@@ -9,6 +9,7 @@ const afterEach = lab.afterEach
 
 const expect = require('code').expect
 const sinon = require('sinon')
+const Promise = require('bluebird')
 
 const Intercom = require('intercom-client')
 const mockIntercom = require('./mocks/mockIntercom')
@@ -63,4 +64,58 @@ describe('orion', () => {
       done()
     })
   }) // end 'constructor'
+
+  describe('nextPage', () => {
+    let orion
+    const samplePageInfo = { page: 1 }
+
+    beforeEach((done) => {
+      orion = new Orion()
+      orion.intercomClient = {
+        nextPage: sinon.stub().returns(Promise.resolve({ body: { page: 2 } }))
+      }
+      Util.canUseIntercom.returns(true)
+      done()
+    })
+
+    it('should do nothing if we cannot use intercom', (done) => {
+      Util.canUseIntercom.reset()
+      Util.canUseIntercom.returns(false)
+      orion.nextPage(samplePageInfo)
+        .then(() => {
+          sinon.assert.calledOnce(Util.canUseIntercom)
+          sinon.assert.notCalled(orion.intercomClient.nextPage)
+        })
+        .asCallback(done)
+    })
+
+    it('should reject if the client is invalid', (done) => {
+      orion.intercomClient = null
+      orion.nextPage(samplePageInfo).asCallback((err) => {
+        expect(err).to.exist()
+        expect(err.message).to.match(/invalid.*client/i)
+        done()
+      })
+    })
+
+    it('should call nextPage on the client', (done) => {
+      orion.nextPage(samplePageInfo)
+        .then(() => {
+          sinon.assert.calledOnce(orion.intercomClient.nextPage)
+          sinon.assert.calledWithExactly(
+            orion.intercomClient.nextPage,
+            samplePageInfo
+          )
+        })
+        .asCallback(done)
+    })
+
+    it('should call nextPage on the client', (done) => {
+      orion.nextPage(samplePageInfo)
+        .then((body) => {
+          expect(body).to.deep.equal({ page: 2 })
+        })
+        .asCallback(done)
+    })
+  })
 }) // end 'orion'
